working audio recorder
