+++ resources/client/components/AudioRecorder/useAudioRecorder.ts
@@ -1,0 +1,128 @@
+// useAudioRecorder.ts
+import { ref, onUnmounted } from "vue";
+
+export function useAudioRecorder(options = { maxDuration: 15 }) {
+  const isRecording = ref(false);
+  const recordingTime = ref(0);
+  const audioBlob = ref<Blob | null>(null);
+  const audioUrl = ref<string | null>(null);
+  const mediaRecorder = ref<MediaRecorder | null>(null);
+  const audioChunks = ref<Blob[]>([]);
+  const timerInterval = ref<number | null>(null);
+  const maxRecordingTime = options.maxDuration;
+
+  // Start recording function
+  const startRecording = async () => {
+    try {
+      // Reset state
+      audioChunks.value = [];
+      recordingTime.value = 0;
+
+      // Request microphone access
+      const stream = await navigator.mediaDevices.getUserMedia({
+        audio: {
+          echoCancellation: true,
+          noiseSuppression: true,
+          autoGainControl: true,
+        },
+      });
+
+      // Create new MediaRecorder
+      mediaRecorder.value = new MediaRecorder(stream);
+
+      // Event handler for data available
+      mediaRecorder.value.ondataavailable = (event) => {
+        if (event.data.size > 0) {
+          audioChunks.value.push(event.data);
+        }
+      };
+
+      // Event handler for when recording stops
+      mediaRecorder.value.onstop = () => {
+        // Create blob from recorded chunks
+        const blob = new Blob(audioChunks.value, { type: "audio/webm" });
+        audioBlob.value = blob;
+
+        // Create URL for audio playback
+        audioUrl.value = URL.createObjectURL(blob);
+
+        // Stop all tracks in the stream to release the microphone
+        stream.getTracks().forEach((track) => track.stop());
+
+        // Clear the timer
+        if (timerInterval.value) {
+          clearInterval(timerInterval.value);
+          timerInterval.value = null;
+        }
+
+        isRecording.value = false;
+      };
+
+      // Start recording
+      mediaRecorder.value.start();
+      isRecording.value = true;
+
+      // Set up timer for recording duration
+      timerInterval.value = window.setInterval(() => {
+        recordingTime.value += 0.1;
+
+        // Auto-stop if max recording time is reached
+        if (recordingTime.value >= maxRecordingTime && isRecording.value) {
+          stopRecording();
+        }
+      }, 100);
+    } catch (error) {
+      console.error("Error starting recording:", error);
+      throw new Error(
+        "Could not access microphone. Please ensure you have granted permission.",
+      );
+    }
+  };
+
+  // Stop recording function
+  const stopRecording = () => {
+    if (mediaRecorder.value && isRecording.value) {
+      mediaRecorder.value.stop();
+    }
+  };
+
+  // Reset recording function
+  const resetRecording = () => {
+    if (audioUrl.value) {
+      URL.revokeObjectURL(audioUrl.value);
+    }
+
+    audioBlob.value = null;
+    audioUrl.value = null;
+    recordingTime.value = 0;
+  };
+
+  // Format time display
+  const formatTime = (time: number): string => {
+    const minutes = Math.floor(time / 60);
+    const seconds = Math.floor(time % 60);
+    return `${minutes}:${seconds.toString().padStart(2, "0")}`;
+  };
+
+  // Clean up on unmount
+  onUnmounted(() => {
+    if (timerInterval.value) {
+      clearInterval(timerInterval.value);
+    }
+
+    if (audioUrl.value) {
+      URL.revokeObjectURL(audioUrl.value);
+    }
+  });
+
+  return {
+    isRecording,
+    recordingTime,
+    audioBlob,
+    audioUrl,
+    startRecording,
+    stopRecording,
+    resetRecording,
+    formatTime,
+  };
+}
